CHROMIUM: drivers: qualcomm/sfe: Flow Acceleration Delay
Normally SFE offloads a flow for fast forwarding as soon as possible.
That's typically immediately after the transport connection is
established and Conntrack / IPTables firewall policy has accepted the
flow as 'confirmed'.  Once SFE offloads the flow Conntrack / IPTables
no longer has any visibility of the packet flow, preventing IPtables
policies from inspecting additional application level protocol
exchanges.

This CL adds a configurable delay before SFE offload to allow
IPTables policies to inspect a limited amount of additional
application level protocol exchange.  The SFE acceleration delay is
expressed in units of 'packets' representing the number of additional
packets that will be delivered to Linux 'slow path' after Conntrack /
IPTables flow is 'confirmed'.  The packet count represents the
number of packets forwarded in either direction for a flow, e.g. if
the SFE flow acceleration delay is set to 10 packets that could be
5 packets in 'orig' direction plus 5 packets in 'reply' direction,
or 10 packets in one direction, or any other combination of 10 packets.

Configuration is exposed via sysfs and is independently controlled
for UDP and TCP protocols.  The default value is '0', which preserves
the existing behavior of SFE offload as soon as possible.  Positive
integer values delay the SFE offload and should be set to the maximum
number of packets required to view application level protocol exchanges
of interest across all IPTables policies.
  /sys/sfe_cm/tcp_flow_accel_delay_pkts
  /sys/sfe_cm/udp_flow_accel_delay_pkts

BUG=b:148971227
TEST=The example below illustrates setting the SFE UDP flow
     acceleration delay to 20 packets.
       % cat /sys/sfe_cm/udp_flow_accel_delay_pkts
       0
       % echo "20" > /sys/sfe_cm/udp_flow_accel_delay_pkts
       % cat /sys/sfe_cm/udp_flow_accel_delay_pkts
       20
     The example below illustrates the SFE flow state for a UDP flow
     when acceleration delay is enabled and 2 packets have been
     forwarded after Conntrack has marked the flow 'confirmed'.
     SFE flow state shows that no packets have been fast forwarded
     ({src,dest}_rx_pkts=0) and the acceleration delay count
     (flow_accel_delay_pkts) has been decremented twice, from 20 to 18.
     Once the acceleration delay count has reached zero SFE will begin
     to fast forward packets, indicated by the {src,dest}_rx_pkts
     incrementing.
       % cat /dev/sfe_ipv4
       <connection protocol="17" src_dev="br-lan" src_ip="192.168.86.25"
         src_port="46407" src_rx_pkts="0" src_rx_bytes="0"
	 dest_dev="wan0" dest_ip="100.90.113.169" dest_port="5001"
	 dest_rx_pkts="0" dest_rx_bytes="0" flow_accel_delay_pkts="18"
       />
TEST=Autotest
     https://jetstream-autotest.corp.google.com/afe/#tab_id=view_job&object_id=1901941
TEST=System testing performed is documented in
     https://docs.google.com/document/d/1-I91VLIC5uSAueJQBmAzqBh5alVIZ6B8F8EqIgdh0B0/edit?usp=sharing

Change-Id: I12b19b1ee76d0a0be5d7cecfd18e5ae3b43bd1c6
Signed-off-by: Danny J. Mitzel <mitzel@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2039952
Reviewed-by: Kan Yan <kyan@chromium.org>